requirements.txt
tqdm
fire
numpy
openai>=2.0.0
python-dotenv
human-eval
pandas>=1.5.0

# CrewAI dependencies (optional, only needed if USE_CREWAI=true)
crewai>=0.1.0
langchain-openai>=0.0.5